Moved admin functionality to actions popup
